Solve la-outer-product

+++ study-plans/math-linear-algebra/la-outer-product/la-outer-product.py
@@ -4,6 +4,6 @@ def outer_product(u, v):
     """
     Returns: float64 matrix of shape (m, n), the outer product u v^T.
     """
-    u = np.array(u, dtype = float)
-    v = np.array(v, dtype = float)
+    u = np.array(u)
+    v = np.array(v)
     return np.outer(u,v)
